loul

+++ commandes/say.js
@@ -1,5 +1,10 @@
 exports.run = (client, message, args) => {
-  const sayMessage = args.join(" ");
-  message.delete().catch(O_o=>{});
-  message.channel.send(sayMessage);
+  if(message.author.id !== "306119836503900161" || message.author.id !== "310474739766394882"){
+   message.channel.send("t'es qui pour me dire quoi dire? xD")
+    return;
+  }else{
+    const sayMessage = args.join(" ");
+    message.delete().catch(O_o=>{});
+    message.channel.send(sayMessage);
+  }
 }
